# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NFT Music",
            "url": "https://nft.threenext.com",
            "email": "seki@threenext.com"
        }
    },
    "paths": {
...
        "/wallets": {
            "get": {
                "description": "仮想通貨のウォレットの情報を出力する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ウォレット情報"
                ],
                "summary": "ウォレットの情報をデータベースから抽出する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.WalletOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "post": {
                "description": "ウォレットアドレスが登録されているかウォレットテーブルを確認し、なければウォレットアドレスをデータベースに格納する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ウォレット情報"
                ],
                "summary": "ウォレットアドレスをデータベースに格納する",
                "parameters": [
                    {
                        "description": "ウォレットアドレス",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.WalletInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.WalletOutput"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ports.CreatedObject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ports.CreatedObject": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "01932564-3432-7b56-b772-7c69586e1897"
                },
                "message": {
                    "type": "string",
                    "example": "データベースには既に存在しています"
                },
                "status_code": {
                    "type": "integer",
                    "example": 201
                }
            }
        },
        "ports.ErrorResponseObject": {
            "type": "object",
            "properties": {
                "error_type": {
                    "type": "string",
                    "example": "リクエストの構文が不正やページが見つからない等のタイプ別"
                },
                "message": {
                    "type": "string",
                    "example": "原因となるのエラーメッセージ"
                },
                "status_code": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "ports.WalletInput": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x29BE6Cd779c7ddbF6cBe3efdeF7Ba6f9CD2C703D"
                }
            }
        },
        "ports.WalletOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x85A4599940764d07f469EdEdAb10a72B727B6669"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-11-07T21:51:36Z"
                },
                "id": {
                    "type": "string",
                    "example": "01932550-9ecc-774f-b1ec-98d3002b9b2e"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}